Refactor: Next.js 구조에 맞게 파일 및 폴더 위치 이동
